move data.file.* to infrastructure
